backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
openai-whisper
librosa
sentence-transformers
numpy
scipy
werkzeug
requests
langchain-ollama
langchain-core
